Refactor file structure & fix action header file
In addition to moving the files around, CMakeLists.txt has been updated
to link the (automatically) generated header file for Motors.action with
the executable node
