+++ jetbot_cpp/include/jetbot_cpp/motor_wrapper.h
@@ -1,5 +1,5 @@
 //#include "action/motor_action.hpp"
-#include "nav2_examples/bt_action_node.hpp"
+#include "jetbot_cpp/nav2_examples/bt_action_node.hpp"
 
 //class MotorAction : public BtActionNode<action::Motor>
 //{
