+++ jetbot_cpp/src/publisher.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 
-#include "publisher.h"
+#include "jetbot_cpp/publisher.h"
 
 /* This example creates a subclass of Node and uses std::bind() to register a
  * member function as a callback from the timer. */
